% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{bfPPalpha}
\alias{bfPPalpha}
\title{Bayes factor for testing power parameter}
\usage{
bfPPalpha(tr, sr, to, so, uv)
}
\arguments{
\item{tr}{Effect estimate of the replication study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{to}{Effect estimate of the original study.}

\item{so}{Standard error of the replication effect estimate.}

\item{uv}{Variance of the unit-information prior.}
}
\value{
Bayes factor (BF > 1 indicates evidence for \eqn{H_0}{H0}, whereas BF
    < 1 indicates evidence for \eqn{H_1}{H1})
}
\description{
This function computes the Bayes factor contrasting \eqn{H_0 :
    \alpha = 0}{H0: alpha = 0} to \eqn{H_0 : \alpha = 1}{H0: alpha = 1} for
    the replication data assuming a normal likelihood. The power parameter
    \eqn{\alpha}{alpha} indicates how much the normal likelihood of the
    original is raised and then incorporated in the power prior for the
    effect size \eqn{\theta}{theta} (e.g. for \eqn{\alpha = 1}{alpha = 1} the
    original data are completely discounted). An initial unit-information
    prior \eqn{\theta \sim \mathrm{N}(0, \mbox{uv})}{theta ~ N(0, uv)} is
    assumed for the effect size \eqn{\theta}{theta} under either hypothesis.
}
\examples{
## use unit variance of 2
bfPPalpha(tr = 0.09,  sr = 0.0518, to = 0.205, so = 0.0506, uv = 2)
}
\seealso{
\code{\link{bfPPtheta}}
}
\author{
Samuel Pawel
}
