% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margLik.R
\name{margLik}
\alias{margLik}
\title{Marginal likelihood of replication effect estimate under power prior model}
\usage{
margLik(tr, to, sr, so, x = 1, y = 1, ...)
}
\arguments{
\item{tr}{Effect estimate of the replication study.}

\item{to}{Effect estimate of the original study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{so}{Standard error of the replication effect estimate.}

\item{x}{Number of successes parameter of beta prior for \eqn{\alpha}{alpha}.
Defaults to 1.}

\item{y}{Number of failures parameter of beta prior for \eqn{\alpha}{alpha}.
Defaults to 1.}

\item{...}{Additional arguments for integration function}
}
\value{
Marginal likelihood
}
\description{
This function computes the marginal likelihood of the
    replication effect estimate \code{tr} under the power prior model
    \deqn{f(\code{tr}|\code{to}, \code{so}, \code{sr}, \code{x}, \code{y}) =
    \int_0^1 \int_{-\infty}^{\infty} \mathrm{N}(\code{tr}; \theta,
    \code{sr}^2) \times \mathrm{N}(\theta; \code{to}, \code{so}^2/\alpha)
    \times \mbox{Beta}(\alpha; \code{x}, \code{y}) ~\mbox{d}\theta~
    \mbox{d}\alpha}{int int N(tr;theta,sr^2) N(theta; to, so^2/alpha)
    Beta(alpha; x, y) dtheta dalpha} using numerical integration.
}
\author{
Samuel Pawel
}
