% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{postPPalpha}
\alias{postPPalpha}
\title{Marginal posterior density of power parameter}
\usage{
postPPalpha(alpha, tr, sr, to, so, x = 1, y = 1, ...)
}
\arguments{
\item{alpha}{Power parameter. Can be a vector.}

\item{tr}{Effect estimate of the replication study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{to}{Effect estimate of the original study.}

\item{so}{Standard error of the replication effect estimate.}

\item{x}{Number of successes parameter of beta prior \eqn{\alpha}{alpha}.
Defaults to 1.}

\item{y}{Number of failures parameter of beta prior \eqn{\alpha}{alpha}.
Defaults to 1.}

\item{...}{Additional arguments for integration function.}
}
\value{
Marginal posterior density of power parameter
}
\description{
This function computes the marginal posterior density of the
    power parameter \eqn{\alpha}{alpha}. A power prior for
    \eqn{\theta}{theta} is constructed by updating an initial flat prior with
    the likelihood of the original data raised to the power of
    \eqn{\alpha}{alpha}. A marginal beta prior \eqn{\alpha \sim
    \mbox{Beta}(x, y)}{alpha ~ Beta(x, y)} is assumed.
}
\examples{
alpha <- seq(0, 1, 0.001)
margpostdens <- postPPalpha(alpha = alpha, tr = 0.1, to = 0.2, sr = 0.05, so = 0.05)
plot(alpha, margpostdens, type = "l", xlab = bquote("Power paramter" ~ alpha),
     ylab = "Marginal posterior density", las = 1)
}
\seealso{
\code{\link{postPP}}, \code{\link{postPPtheta}}
}
\author{
Samuel Pawel
}
