% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{postPPtheta}
\alias{postPPtheta}
\title{Marginal posterior density of effect size}
\usage{
postPPtheta(theta, tr, sr, to, so, x = 1, y = 1, alpha = NA, ...)
}
\arguments{
\item{theta}{Effect size. Can be a vector.}

\item{tr}{Effect estimate of the replication study.}

\item{sr}{Standard error of the replication effect estimate.}

\item{to}{Effect estimate of the original study.}

\item{so}{Standard error of the replication effect estimate.}

\item{x}{Number of successes parameter \eqn{x}{x} for beta prior of power
parameter \eqn{\alpha}{alpha}. Defaults to 1. Is only taken into account
when \code{alpha = NA}.}

\item{y}{Number of failures parameter \eqn{y}{y} for beta prior of power
parameter \eqn{\alpha}{alpha}. Defaults to 1. Is only taken into account
when \code{alpha = NA}.}

\item{alpha}{Power parameter. Can be set to a number between 0 and 1.
Defaults to \code{NA}.}

\item{...}{Additional arguments for integration function.}
}
\value{
Marginal posterior density of effect size
}
\description{
This function computes the marginal posterior density of the
    effect size \eqn{\theta}{theta}. A power prior for \eqn{\theta}{theta} is
    constructed by updating an initial flat prior with likelihood of the
    original data raised to the power of \eqn{\alpha}{alpha}. The power
    parameter \eqn{\alpha}{alpha} can either be fixed to some value between 0
    and 1, or it can have a Beta distribution \eqn{\alpha \sim
    \mbox{Beta}(\code{x}, \code{y})}{alpha ~ Beta(\code{x}, \code{y})}.
}
\examples{
theta <- seq(0, 0.6, 0.001)
margpostdens <- postPPtheta(theta = theta, tr = 0.1, to = 0.2, sr = 0.05, so = 0.05)
plot(theta, margpostdens, type = "l", xlab = bquote("Effect size" ~ theta),
     ylab = "Marginal posterior density", las = 1)
}
\seealso{
\code{\link{postPP}}, \code{\link{postPPalpha}}
}
\author{
Samuel Pawel
}
