% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppRepPKG.R
\docType{package}
\name{ppRepPKG}
\alias{ppRepPKG}
\title{ppRep: Bayesian design and analysis of replication studies using power
    priors}
\description{
The ppRep package features methods for design and analysis of
    replication studies using power priors assuming normality. Power priors
    are a useful class of informative prior distributions that allow to
    incorporate the data from the original study in the analysis of the
    replication data.
}
\references{
Pawel, S., .... (2022). Power priors for design and analysis of
    replication studies
}
